Fix db scheme

+++ interchain-indexer/interchain-indexer-entity/src/codegen/indexer_failures.rs
@@ -7,7 +7,7 @@ use sea_orm::entity::prelude::*;
 pub struct Model {
     #[sea_orm(primary_key)]
     pub id: i64,
-    pub bridge_id: i64,
+    pub bridge_id: i32,
     pub chain_id: i64,
     pub from_block: i64,
     pub to_block: i64,
